Add previous results as initial guesses to frechet.jl
Results from a previous parameterization can be loaded by setting 'previous = 1' in line 6 of frechet.jl.
The load command is in a single location (line 71). Certain statements that are needed when 'previous = 0' are sprinkled around the script; we need to clean this up at some point.
Results are now stored in ./parameterization subdirectory.