Change guess for labor market barriers in calibrate_tau to zeroes in frechet.jl
The optimization using calibrate_tau based on a guess from a previous parameterization was drifting towards barriers in excess of 100%. It is still not clear why this happens. An uneducated guess of tau=0 for women in all occupations is more time-consuming but finds a solution that minimizes the sum of squared differences between model and data occupational shares.

A separate issue is related to productivity growth and the definition of 'delta' in the Julia script. For now, delta is commented. This, however, implies that 'convG' does not converge. This needs to be resolved.